requirements.txt
# Deep Learning Frameworks (Google Colab Compatible)
# For CUDA support: pip install torch==2.3.0 torchvision==0.18.0 --index-url https://download.pytorch.org/whl/cu118
torch==2.3.0
torchvision==0.18.0
ultralytics==8.1.25

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Model Training & Monitoring
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Data Download & Management
kaggle>=1.5.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
click>=8.1.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
jupyter>=1.0.0

# Deployment
gunicorn>=21.0.0
docker>=6.1.0